fullRestructureOfProj
